emulator/pyproject.toml
[project]
name = "scouterhud-emulator"
version = "0.1.0"
description = "QR-Link Device Emulator Hub for ScouterHUD"
requires-python = ">=3.10"
dependencies = [
    "paho-mqtt>=2.0",
    "pyyaml>=6.0",
    "qrcode[pil]>=7.0",
    "reportlab>=4.0",
    "Pillow>=10.0",
]

[project.scripts]
scouterhud-emulator = "emulator:main"
